Fix bearing calculation. Change type degree int to double

+++ base-domain/src/main/java/com/example/basedomain/dto/BearingPosition.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class BearingPosition {
-    private int degree;
+    private double degree;
     private double distance;
     private String sourceName;
